backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
beautifulsoup4>=4.12.2
requests>=2.31.0
# nltk>=3.8.1 # Removed as likely unused
python-dotenv>=1.0.0
pydantic>=2.4.2
python-multipart>=0.0.6
google-generativeai>=0.3.1
networkx>=3.1
markdown-it-py
playwright>=1.41.0
tenacity>=8.2.3 # For automatic retries

# After installing requirements, run:
# python -m playwright install chromium 